Playing with how Roles and Permissions interact.
We acknowledge that roles/permissions should have time gates for testing or emergency purposes, but we're not there yet. It's likely we'll have to implement another layer of abstraction to the system to get there, but what that looks like is still about as clear as silt.

The other thing we acknowledge is that while MOST things will only care about user-based permissions in order to be edited (which has been represented by a "permissionRequiredToEdit" field in some entities), some things (such as groups) can care about membership-based permissions as well. We do not have a system in place yet to set and grant default permissions for these gates, but they're on their way.